apps/backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.2.0
redis==5.0.3
celery==5.3.6
psycopg2-binary>=2.9.11
pytest-django==4.8.0
locust==2.24.0
daphne==4.1.0
websockets==12.0
websocket-client==1.7.0
dj-database-url==2.1.0
django-cors-headers==4.3.1
gevent==23.9.1
structlog
whitenoise==6.6.0
gunicorn==21.2.0
django-celery-beat==2.6.0